Temporary removed the new buss implementation and changed back to the old way of sending sensordata to the raspberry. Also changed to send negative values for roll/pitch/yaw

+++ Arduino/src/Mainc/SensorDataStruct.h
@@ -37,6 +37,7 @@
 #define CELL2 1
 #define CELL3 2
 
+
 struct ps3Controller {
   int button[NUM_OF_BUTTONS];
   int stick1_x;
@@ -48,10 +49,10 @@ struct ps3Controller {
 struct sensordata {
   /*Angles*/
   int motorSpeed[4];
-  float cellVoltage[3];
-  float angleYaw;
-  float anglePitch;
-  float angleRoll;
+  int cellVoltage[3];
+  int angleYaw;
+  int anglePitch;
+  int angleRoll;
   /*Accelerations*/
   VectorInt16 acc;
